+++ src/modinfodialogconflicts.cpp
@@ -527,7 +527,7 @@ void ConflictsTab::openItems(QTreeView* tree)
   // the menu item is only shown for a single selection, but handle all of them
   // in case this changes
   for_each_in_selection(tree, [&](const ConflictItem* item) {
-    core().runFile(parentWidget(), item->fileName());
+    core().processRunner().runFile(parentWidget(), item->fileName());
     return true;
   });
 }
